+++ c3d_math/distances.c
@@ -6,7 +6,7 @@
 /*   By: zanikin <zanikin@student.42yerevan.am>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/23 18:07:22 by zanikin           #+#    #+#             */
-/*   Updated: 2024/10/03 16:06:24 by zanikin          ###   ########.fr       */
+/*   Updated: 2024/10/09 20:00:10 by zanikin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,17 +14,17 @@
 
 #include "t_vec.h"
 
-float	square_distance(const t_vec *p1, const t_vec *p2)
+double	square_distance(const t_vec *p1, const t_vec *p2)
 {
-	float	x;
-	float	y;
+	double	x;
+	double	y;
 
 	x = p1->x - p2->x;
 	y = p1->y - p2->y;
 	return (x * x + y * y);
 }
 
-float	distance(const t_vec *p1, const t_vec *p2)
+double	distance(const t_vec *p1, const t_vec *p2)
 {
 	return (sqrtf(square_distance(p1, p2)));
 }
