+++ c3d_math/set.c
@@ -6,14 +6,14 @@
 /*   By: zanikin <zanikin@student.42yerevan.am>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/21 22:27:40 by zanikin           #+#    #+#             */
-/*   Updated: 2024/10/03 16:06:37 by zanikin          ###   ########.fr       */
+/*   Updated: 2024/10/09 20:00:11 by zanikin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "t_vec.h"
 #include "t_ivec.h"
 
-void	set_vec(t_vec *v, float x, float y)
+void	set_vec(t_vec *v, double x, double y)
 {
 	v->x = x;
 	v->y = y;
